refactor/organiza arquitetura da aplicacao
 Foi organizada a arquitetura da aplicação, no diretório MediaVirtu/src/javascript, separando cada código javascript por responsabilidades.
 No diretório modificado, teremos os arquivos que importam das pastas nesse diretório, e exportam para o main.js; E o main.js, que irá impordar esses arquivos no mesmo diretório, e arquivos simples como os .json da pasta data, e o construtor do preloader da página, na pasta componentes simples.
 Em componentes_simples, teremos todos os arquivos que apenas retornam estrutura html simples e não importam nenhum outro arquivo. Qualquer outro arquivo de outra pasta que constroi html, pode importar essas funções.
 Em eventos, teremos todos os arquivos que contém funções que criam eventos. Essas funçoes serao importadas para ativaEventos.js
 Em inicio, teremos todos os arquivos contendo todas as funções necessárias para a construção da página inicial. apenas inicial.js exporta todo html produzido aqui para fora.
