+++ MediaVirtu/src/estilos/corpo_principal.css
@@ -46,15 +46,15 @@
     width: 100%;
     display: flex;
     flex-direction: column;
+    justify-content: center;
     align-items: center;
     background-color: black;
     color: white;
     z-index: 2;
-    transition: height 0.5s ease-in-out;
+    transition: height 1s ease-in-out;
     overflow: hidden;
-    height: 0px; /*Valor a ser alterado para exibir ou remover preloader*/
 }
 
 #preloader h1 {
-    font-size: 70px;
+    font-size: 8vw;
 }
