+++ MediaVirtu/src/javascript/add_conteudo.js
@@ -1,0 +1,6 @@
+import pagina from "./pagina.js";
+
+export default function add_conteudo () {
+    const conteudo_pagina = document.querySelector("#corpo-principal");
+    conteudo_pagina.innerHTML = pagina(localStorage.getItem("pagina_atual"));
+}
