+++ MediaVirtu/src/javascript/componentes_simples/gerador_aleatorio.js
