+++ MediaVirtu/src/javascript/componentes_simples/linha_divisora.js
