+++ MediaVirtu/src/javascript/main.js
@@ -1,68 +1,36 @@
-import cabecalho from "./cabecalho.js"
-import barrasNavegacaoLateral from "./barras_navegacao_lateral.js"
-import pagina from "./corpo_principal.js"
+import cabecalho from "./cabecalho.js";
+import barrasNavegacaoLateral from "./barras_navegacao_lateral.js";
+import preloader from "./componentes_simples/preloader_construtor.js";
+import add_conteudo from "./add_conteudo.js";
+import ativaEventos from "./ativaEventos.js";
 import banco_posts from "../data/shitposts.json" assert { type: "json" };
 import banco_comentarios from "../data/comentarios.json" assert { type: "json" };
-import addLikesDislikesEventos from "./likes_dislikes.js"
-import addComentarioEvento from "./comentario_evento.js"
 
 
 localStorage.setItem("banco_posts", JSON.stringify(banco_posts));
 
+if (!localStorage.getItem("lista_likes")) localStorage.setItem("lista_likes", []);
+if (!localStorage.getItem("lista_dislikes")) localStorage.setItem("lista_dislikes", []);
+if (!localStorage.getItem("banco_comentarios")) localStorage.setItem("banco_comentarios", JSON.stringify(banco_comentarios));
+
 localStorage.setItem("pagina_atual", "Inicio");
 
-if (!localStorage.getItem("lista_likes")) localStorage.setItem("lista_likes", []);
-
-if (!localStorage.getItem("lista_dislikes")) localStorage.setItem("lista_dislikes", []);
-
-if (!localStorage.getItem("banco_comentarios")) localStorage.setItem("banco_comentarios", JSON.stringify(banco_comentarios));
 
 const root = document.querySelector("#root");
 
 
+inicializar();
+add_conteudo();
+ativaEventos();
+
+
+
 function inicializar () {
     root.innerHTML = `
+    ${ preloader() }
     ${ cabecalho() }
     ${ barrasNavegacaoLateral() }
     <main id = "corpo-principal"></main>
     `;
     
-  }
-  
-inicializar();
-add_conteudo()
-eventosBarrasNavegacaoLateral();
-addLikesDislikesEventos()
-addComentarioEvento()
-  
-
-function add_conteudo () {
-    const conteudo_pagina = document.querySelector("#corpo-principal");
-    conteudo_pagina.innerHTML = pagina(localStorage.getItem("pagina_atual"));
 }
-
-
-function eventosBarrasNavegacaoLateral() {
-  const links = document.querySelectorAll("[data-page]");
-  links.forEach((link) => {
-    link.addEventListener("click", (e) => {
-      e.preventDefault();
-
-      const page = e.currentTarget.dataset.page;
-
-      localStorage.setItem("pagina_atual", page);
-
-      add_conteudo();
-    });
-  });
-}
-
-/* Permite voltar para a "aba anterior"...
-Evita do usuário apertar no botão de voltar e acabar saindo do site. */
-window.addEventListener("popstate", (e) => {
-  if (e.state && e.state.page) {
-    localStorage.setItem("pagina_atual", e.state.page);
-    
-    render();
-  }
-});
